Added ability to see node_modules or exit

+++ nm_remove.py
@@ -1,8 +1,27 @@
 import os
 
-os.system("find . -name 'node_modules' -type d -prune")
-# '{}' is a placeholder which find replaces with the file path it found, while + tells find to append all the file paths to a single command, rather than running rm for each
+def init(): 
+    sendIt = int(input('See node_modules first? \n 1 - Yes, please!  \n 2 - No, send it \n 3 - What have I done?? Get me out of here \n'))
 
-os.system("find . -name 'node_modules' -type d -prune -exec rm -rf '{}' +")
+    if sendIt == 1:
+        print("You didn't send it :(")
+        os.system("find . -name 'node_modules' -type d -prune")
 
-print("node_module pruning complete!!")
+    elif sendIt == 2:
+        print("Are you silly? I'm still gonna send it")
+        os.system("find . -name 'node_modules' -type d -prune")
+
+        # '{}' is a placeholder which find replaces with the file path it found, while + tells find to append all the file paths to a single command, rather than running rm for each
+        os.system("find . -name 'node_modules' -type d -prune -exec rm -rf '{}' +")
+
+        print("node_module pruning complete!!")
+
+    elif sendIt ==3:
+        return
+
+    else:
+        print('You had 3 valid options, choose one')
+        init()
+
+
+init()
